updated sh

+++ train_single.sh
@@ -3,14 +3,18 @@
 datamode="noniid_max_single"
 dataset="stanford"
 mode="per_study"
-desc="no_positive_weight_layernorm"
+desc="no_positive_weight_dropout"
 runtime_dir="20190927_${datamode}_${dataset}_${mode}_${desc}"
+start_epoch=1
 
-rm -rf runtime/$runtime_dir
+if [ $start_epoch -gt 1 ]; then
+  rm -rf runtime/$runtime_dir
+fi
 
 python train_noniid.py \
   --cuda 0 \
   --main-dataset $dataset \
   --runtime-dir $runtime_dir \
+  --start-epoch $start_epoch \
   --tensorboard \
   --ignore-repo-dirty \
